browser_agent/browser_agent_system/requirements.txt
google-generativeai>=0.8.0
playwright>=1.40.0
python-dotenv>=1.0.0
colorama>=0.4.6
